+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -17,6 +17,12 @@ import 'feed/repost.dart';
 import 'feed/reposted_by.dart';
 import 'feed/thread.dart';
 import 'feed/timeline.dart';
+import 'graph/follow.dart';
+import 'graph/followers.dart';
+import 'graph/follows.dart';
+import 'graph/mute.dart';
+import 'graph/mutes.dart';
+import 'graph/unmute.dart';
 import 'notification/notification_count.dart';
 import 'notification/notifications.dart';
 import 'notification/seen_notifications.dart';
@@ -49,6 +55,15 @@ final notificationCommands = <BskyCommand>[
   NotificationsCommand(),
   NotificationCountCommand(),
   SeenNotificationsCommand(),
+];
+
+final graphCommands = <BskyCommand>[
+  FollowCommand(),
+  FollowsCommand(),
+  FollowersCommand(),
+  MuteCommand(),
+  UnmuteCommand(),
+  MutesCommand(),
 ];
 
 final unspeccedCommands = <BskyCommand>[