+++ packages/bluesky_cli/lib/src/runner/bsky_runner.dart
@@ -3,6 +3,8 @@
 // modification, are permitted provided the conditions.
 
 import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
 
 import 'package:cli_util/cli_logging.dart';
 import 'package:xrpc/xrpc.dart';
@@ -23,13 +25,24 @@ class Bsky extends _Bsky {
 
   @override
   Future<void> run() async {
-    final response = await action.call();
-    final status = response.status;
+    try {
+      final response = await action.call();
+      final status = response.status;
 
-    logger
-      ..success('${status.code} ${status.message}')
-      ..success(response.request.toString())
-      ..success(response.data);
+      logger
+        ..success('${status.code} ${status.message}')
+        ..success(response.request.toString())
+        ..success(response.data);
+    } on XRPCException catch (e) {
+      final status = e.response.status;
+
+      logger
+        ..error('${status.code} ${status.message}')
+        ..error(e.response.request.toString())
+        ..error(jsonEncode(e.response.data.toJson()));
+
+      exitCode = 1;
+    }
   }
 }